+++ NEWS.md
@@ -1,6 +1,6 @@
 Phan NEWS
 
-?? ??? 2019, Phan 2.0.1 (dev)
+01 Jun 2019, Phan 2.1.0
 -----------------------
 
 New features(CLI, Configs):
@@ -22,7 +22,7 @@ New features(Analysis):
   e.g. `static $a = null; static $b = null; if ($a === null) { $a = $b = rand(0,10); } use($a, $b)`
   will now also infer that $b is non-null.
 + Infer from `return new static();` and `return $this;` that the return type of a method is `@return static`, not `@return self` (#2797)
-  (and propogate that to inherited methods)
+  (and propagate that to inherited methods)
 + Fix some false positives when casting array types containing `static` to types containing the class or its ancestors. (#2797)
 + Add `PhanTypeInstantiateAbstractStatic` and `PhanTypeInstantiateTraitStaticOrSelf` as lower-severity warnings about `return new self()` and `return new static()` (#2797)
   (emitted in static methods of abstract classes)